+++ modules/autotagging-commons/src/test/java/org/opensearch/rule/InMemoryRuleProcessingServiceTests.java
@@ -17,10 +17,13 @@ import org.opensearch.rule.storage.AttributeValueStoreFactory;
 import org.opensearch.rule.storage.DefaultAttributeValueStore;
 import org.opensearch.test.OpenSearchTestCase;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+
+import static org.opensearch.rule.attribute_extractor.AttributeExtractor.CombinationStyle.AND;
 
 public class InMemoryRuleProcessingServiceTests extends OpenSearchTestCase {
     InMemoryRuleProcessingService sut;
@@ -31,7 +34,7 @@ public class InMemoryRuleProcessingServiceTests extends OpenSearchTestCase {
             WLMFeatureType.WLM,
             DefaultAttributeValueStore::new
         );
-        sut = new InMemoryRuleProcessingService(attributeValueStoreFactory);
+        sut = new InMemoryRuleProcessingService(attributeValueStoreFactory, WLMFeatureType.WLM.getPrioritizedAttributesList());
     }
 
     public void testAdd() {
@@ -122,7 +125,8 @@ public class InMemoryRuleProcessingServiceTests extends OpenSearchTestCase {
     }
 
     private static List<AttributeExtractor<String>> getAttributeExtractors(List<String> extractedAttributes) {
-        List<AttributeExtractor<String>> extractors = List.of(new AttributeExtractor<String>() {
+        List<AttributeExtractor<String>> extractors = new ArrayList<>();
+        extractors.add(new AttributeExtractor<String>() {
             @Override
             public Attribute getAttribute() {
                 return TestAttribute.TEST_ATTRIBUTE;
@@ -131,6 +135,11 @@ public class InMemoryRuleProcessingServiceTests extends OpenSearchTestCase {
             @Override
             public Iterable<String> extract() {
                 return extractedAttributes;
+            }
+
+            @Override
+            public CombinationStyle getCombinationStyle() {
+                return AND;
             }
         });
         return extractors;
@@ -151,6 +160,11 @@ public class InMemoryRuleProcessingServiceTests extends OpenSearchTestCase {
         @Override
         public Map<String, Attribute> getAllowedAttributesRegistry() {
             return Map.of("test_attribute", TestAttribute.TEST_ATTRIBUTE);
+        }
+
+        @Override
+        public List<Attribute> getPrioritizedAttributesList() {
+            return List.of(TestAttribute.TEST_ATTRIBUTE);
         }
     }